chore: Update Menu.stories.js to Menu.stories.tsx and add MenuProps type definition
